django/requirements.txt
Django==3.2.25
django_import_export==3.2.0
djangorestframework==3.15.1
PyJWT==2.8.0
PyJWT==2.8.0
Requests==2.32.2
